+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- MS3Record.samprate() consistently returns the rate in samples/second,
+  and MS3Record.samplrate_raw() returns the record's raw value.
+
 
 ## [0.0.4] - 2025-9-24
 ### Fixed